[pyOpenMS] Require nanobind 3 and py-build-cmake 0.5 (#10133, PR 1) (#10138)
* [pyOpenMS] Require nanobind 3 and py-build-cmake 0.5 (#10133, PR 1)

Establishes the nanobind 3 baseline in normal linked mode, ahead of the
split-mode/abi3 wheel change. Wheel production is unchanged: one build per
supported Python version, no BACKEND_MODULE or STABLE_ABI_VERSION yet.

Build metadata (src/pyOpenMS/pyproject.toml), in both [build-system].requires
and [dependency-groups].build:
- nanobind>=2.4.0 -> nanobind>=3.0.1,<4. The declared range no longer lets an
  isolated build resolve a nanobind the sources cannot compile against, and the
  upper bound keeps a future nanobind 4 out.
- py-build-cmake>=0.3.0 -> py-build-cmake~=0.5.1, so the declared minimum covers
  the wheel ABI configuration the planned cp311-abi3 wheels need.

Type casters: nanobind 3 widened the custom-caster from_python() flags from
uint8_t to uint32_t. All 11 signatures across the five caster headers now take
uint32_t, matching the protocol nanobind 3 documents in nb_cast.h.

CMake (src/pyOpenMS/CMakeLists.txt):
- The FetchContent fallback fetches nanobind v3.0.1 instead of v2.10.0 (which
  was also yanked upstream for a CMake regression).
- find_package(nanobind) now requests 3.0.1. nanobind ships a SameMajorVersion
  config, so this accepts >= 3.0.1 and < 4; the previous unversioned call would
  happily have selected an installed 2.x.
- A nanobind supplied by an enclosing project is checked rather than bypassing
  the requirement, and a caller-supplied nanobind_ROOT/nanobind_DIR is no longer
  overwritten by the interpreter probe, only version-checked like any other.
- The probe result is checked, and the selected nanobind version and directory
  are reported in the configure log and the configuration summary.
- Stub generation imports nanobind from ${Python_EXECUTABLE}; configure now warns
  when that interpreter has no nanobind, or one older than the frontend the
  modules are compiled against.
- The MSVC __forceinline patch is kept: nanobind 3.0.1 still has it in nb_defs.h.

Docs/changelog: the conda quick-start created a python=3.10 environment, which
is no longer a supported version; the dependency list and the PyOpenMS section
of the 3.6.0 changelog record the new nanobind minimum.

Verified locally against nanobind 3.0.1 (the current 3.x release, so also what
CI resolves):
- All 15 binding translation units (13 domain modules, main_module,
  arrow_zerocopy) compile clean.
- A module built from the real caster headers gives byte-identical results on 26
  str/bytes/list/tuple/set/dict conversion and rejection probes under
  (old casters, nanobind 2.9.2) and (new casters, nanobind 3.0.1).
- nanobind.stubgen 3.0.1 emits the same ndarray/None annotations as 2.x for the
  nb::ndarray shapes pyOpenMS uses, so fix_stubs.py and stubs.patch still apply.
- Discovery exercised for: nanobind 3 installed, nanobind 2.x installed (rejected,
  falls back to fetching v3.0.1), no nanobind installed, and nanobind 3 / 2.x
  supplied by an enclosing project (accepted / fatal error).


* [pyOpenMS] Compare the full nanobind version of a supplied installation

The enclosing-project branch accepted any nanobind whose headers contain
nb_backend.h. That header is also present in nanobind 3.0.0, so a project
supplying 3.0.0 passed a check that is meant to require >= 3.0.1, and the
reported version was only ever "3 (exact version unknown)".

nanobind.h defines NB_VERSION_MAJOR / NB_VERSION_MINOR / NB_VERSION_PATCH, so
read those and compare the complete version instead. A supplied installation
whose version cannot be determined is now rejected rather than accepted, with a
message pointing at find_package(nanobind 3.0.1 CONFIG) as the way to make the
version known. The components are collected and counted rather than tested for
truth, because 0 is a legitimate value for any of them (3.0.1 itself has
NB_VERSION_MINOR 0).

Found by CodeRabbit review on #10138.

Verified against the same harness as the original block:
- nanobind 3.0.1 supplied without nanobind_VERSION -> parsed as 3.0.1 from the
  header and accepted;
- a tree carrying NB_VERSION_PATCH 0 (i.e. 3.0.0) and nb_backend.h -> rejected,
  "the enclosing project supplied nanobind 3.0.0", which is exactly the case the
  previous check let through;
- NB_DIR without nanobind.h -> rejected as undeterminable;
- no regression on the other paths: plain find_package discovery and an
  enclosing project that does export nanobind_VERSION both still report 3.0.1,
  and a full OpenMS configure with PYOPENMS=ON still selects nanobind 3.0.1.


---------

+++ src/pyOpenMS/bindings/type_casters/openms_datavalue_caster.h
@@ -54,7 +54,7 @@ public:
                    const_name("None | int | float | str | bytes | "
                              "list[str] | list[int] | list[float]"))
 
-    bool from_python(handle src, uint8_t flags, cleanup_list* cleanup) noexcept {
+    bool from_python(handle src, uint32_t flags, cleanup_list* cleanup) noexcept {
         // Handle None -> Empty DataValue
         if (src.is_none()) {
             value = OpenMS::DataValue();
@@ -314,7 +314,7 @@ public:
                    const_name("None | int | float | str | bytes | "
                              "list[str] | list[int] | list[float]"))
 
-    bool from_python(handle src, uint8_t flags, cleanup_list* cleanup) noexcept {
+    bool from_python(handle src, uint32_t flags, cleanup_list* cleanup) noexcept {
         // Handle None -> Empty ParamValue
         if (src.is_none()) {
             value = OpenMS::ParamValue();
